added api and web projects

+++ src/Business/MasGlobal.Employees.Business/Model/HourlyEmployee.cs
@@ -2,6 +2,8 @@
 {
     public class HourlyEmployee : Employee
     {
-        public override int AnnualSalary => 120 * base.Salary * 12;
+        // TODO: consider making these constants configurable.
+
+        public override double AnnualSalary => 120 * base.Salary * 12;
     }
 }
